requirements.txt
astropy
ccdproc
matplotlib
numpy
photutils
pandas
scipy
pyqt=4